+++ influx/influx.go
@@ -2,18 +2,17 @@ package influx
 
 import (
 	"context"
-	"crypto/tls"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io/ioutil"
-	"net"
 	"net/http"
 	"net/url"
 	"strings"
 	"time"
 
 	"github.com/influxdata/chronograf"
+	"github.com/influxdata/chronograf/util"
 )
 
 var _ chronograf.TimeSeries = &Client{}
@@ -24,35 +23,6 @@ var _ chronograf.Databases = &Client{}
 var skipVerifyTransport *http.Transport
 var defaultTransport *http.Transport
 
-// CreateTransport create a new transport
-func CreateTransport(skipVerify bool) *http.Transport {
-	var transport *http.Transport
-	if cloneable, ok := http.DefaultTransport.(interface{ Clone() *http.Transport }); ok {
-		transport = cloneable.Clone() // available since go1.13
-	} else {
-		// This uses the same values as http.DefaultTransport
-		transport = &http.Transport{
-			Proxy: http.ProxyFromEnvironment,
-			DialContext: (&net.Dialer{
-				Timeout:   30 * time.Second,
-				KeepAlive: 30 * time.Second,
-				DualStack: true,
-			}).DialContext,
-			MaxIdleConns:          100,
-			IdleConnTimeout:       90 * time.Second,
-			TLSHandshakeTimeout:   10 * time.Second,
-			ExpectContinueTimeout: 1 * time.Second,
-		}
-	}
-	if skipVerify {
-		if transport.TLSClientConfig == nil {
-			transport.TLSClientConfig = &tls.Config{}
-		}
-		transport.TLSClientConfig.InsecureSkipVerify = true
-	}
-	return transport
-}
-
 // SharedTransport returns a shared transport with requested TLS InsecureSkipVerify value
 func SharedTransport(skipVerify bool) *http.Transport {
 	if skipVerify {
@@ -62,8 +32,8 @@ func SharedTransport(skipVerify bool) *http.Transport {
 }
 
 func init() {
-	skipVerifyTransport = CreateTransport(true)
-	defaultTransport = CreateTransport(false)
+	skipVerifyTransport = util.CreateTransport(true)
+	defaultTransport = util.CreateTransport(false)
 }
 
 // Client is a device for retrieving time series data from an InfluxDB instance