+++ server/server.go
@@ -33,6 +33,7 @@ import (
 	clog "github.com/influxdata/chronograf/log"
 	"github.com/influxdata/chronograf/oauth2"
 	"github.com/influxdata/chronograf/server/config"
+	"github.com/influxdata/chronograf/util"
 	client "github.com/influxdata/usage-client/v1"
 	flags "github.com/jessevdk/go-flags"
 )
@@ -692,14 +693,10 @@ func (s *Server) Serve(ctx context.Context) {
 		return
 	}
 
-	s.oauthClient = http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{
-				InsecureSkipVerify: s.GenericInsecure,
-				RootCAs:            certs,
-			},
-		},
-	}
+	transport := util.CreateTransport(true)
+	transport.TLSClientConfig.InsecureSkipVerify = s.GenericInsecure
+	transport.TLSClientConfig.RootCAs = certs
+	s.oauthClient = http.Client{Transport: transport}
 
 	auth := oauth2.NewCookieJWT(s.TokenSecret, s.AuthDuration, s.InactivityDuration)
 	providerFuncs := []func(func(oauth2.Provider, oauth2.Mux)){